refactor: used barrel import for context and minor changes
