+++ src/Context/index.js
@@ -1,0 +1,3 @@
+import { useAuth } from "./AuthContext/AuthContext";
+import { useUser } from "./UserContext/UserContext";
+export { useAuth, useUser };
